% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_documentation.R
\docType{data}
\name{seismic_bumps}
\alias{seismic_bumps}
\title{Seismoacoustic data form coal mine}
\format{A data frame with 2584 rows and 19 variables}
\description{
The data describes the problem of high energy (higher than 10^4 J) seismic bumps
forecasting in a coal mine. Data comes from two of longwalls located in a Polish coal mine
}
\details{
Attribute information.
\itemize{
 \item{seismic} {result of shift seismic hazard assessment in the mine working obtained by the seismic
 method (a - lack of hazard, b - low hazard, c - high hazard, d - danger state)}
 \item{seismoacoustic}{result of shift seismic hazard assessment in the mine working obtained by the
 seismoacoustic method}
 \item{shift}{information about type of a shift (W - coal-getting, N -preparation shift)}
 \item{genergy}{seismic energy recorded within previous shift by the most active geophone (GMax) out of
 geophones monitoring the longwall}
 \item{gpuls}{a number of pulses recorded within previous shift by GMax}
 \item{gdenergy}{ a deviation of energy recorded within previous shift by GMax from average energy recorded
 during eight previous shifts}
 \item{gdpuls}{a deviation of a number of pulses recorded within previous shift by GMax from average number
 of pulses recorded during eight previous shifts}
 \item{ghazard}{result of shift seismic hazard assessment in the mine working obtained by the
 seismoacoustic method based on registration coming form GMax only}
 \item{nbumps}{the number of seismic bumps recorded within previous shift}
 \item{nbumps2}{the number of seismic bumps (in energy range [0^2,10^3)) registered within previous shift}
 \item{nbumps3}{the number of seismic bumps (in energy range [10^3,10^4)) registered within previous shift}
 \item{nbumps4}{the number of seismic bumps (in energy range [10^4,10^5)) registered within previous shift}
 \item{nbumps5}{the number of seismic bumps (in energy range [10^5,10^6)) registered within previous shift}
 \item{nbumps6}{the number of seismic bumps (in energy range [10^6,10^7)) registered within previous shift}
 \item{nbumps7}{the number of seismic bumps (in energy range [10^7,10^8)) registered within previous shift}
 \item{nbumps89}{the number of seismic bumps (in energy range [10^8,10^10)) registered within previous shift}
 \item{energy}{total energy of seismic bumps registered within previous shift}
 \item{maxenergy}{the maximum energy of the seismic bumps registered within previous shift}
 \item{class}{the decision attribute - "1" means that high energy seismic bump occurred in the next shift
 ("hazardous state"), "0" means that no high energy seismic bumps occurred in the next shift
 ("non-hazardous state").}}
}
\keyword{datasets}
